refactor(plugin): move plugin classes into a new top-level package
