+++ server/src/main/kotlin/com/github/bedrockecs/plugin/zimpl/Log.kt
@@ -1,4 +1,4 @@
-package com.github.bedrockecs.server.plugin.zimpl
+package com.github.bedrockecs.plugin.zimpl
 
 import org.slf4j.LoggerFactory
